Rename worldstate, db, updateset as util tools
