+++ cmd/util-db/db/validator.go
