+++ cmd/util-worldstate/state/info.go
